엔진 핵심 기능 개발 > 기초 렌더링 파이프라인 구성
- 화면 단색 초기화 기능까지 개발

+++ Engine/Runtime/Game/Private/D3D12RHI/D3D12OfflineDescriptorManager.cpp
@@ -11,7 +11,7 @@ using namespace SC::Runtime::Core;
 using namespace SC::Runtime::Game::D3D12RHI;
 using namespace std;
 
-D3D12OfflineDescriptorManager::D3D12OfflineDescriptorManager(ID3D12Device* device, D3D12_DESCRIPTOR_HEAP_TYPE type, uint32 numDescriptorsPerAllocator) : Super()
+D3D12OfflineDescriptorManager::D3D12OfflineDescriptorManager(ID3D12Device* device, D3D12_DESCRIPTOR_HEAP_TYPE type, size_t numDescriptorsPerAllocator) : Super()
 	, Type(type)
 	, NumDescriptorsPerAllocator(numDescriptorsPerAllocator)
 	, device(nullptr)
